+++ src/main/java/com/astartes/ultramar/exception/UltramarineUpdateException.java
@@ -2,7 +2,7 @@ package com.astartes.ultramar.exception;
 
 // Exception pour les erreurs de mise à jour
 public class UltramarineUpdateException extends RuntimeException {
-    public UltramarineUpdateException(int id) {
+    public UltramarineUpdateException(Long id) {
         super("Failed to update Ultramarine with id " + id);
     }
 }
